requirements.txt
# corresponds to production image dependencies https://gitlab.adsrvr.org/thetradedesk/teams/dataproc/airflow-ansible/-/blob/master/airflow.yaml#L26
apache-airflow[async,amazon,celery,cncf.kubernetes,docker,elasticsearch,google,http,hashicorp,microsoft.azure,microsoft.mssql,postgres,redis,ssh,statsd,tableau,vertica,virtualenv,providers-common-sql,sftp]==2.10.5
wtforms==3.1.2
flask-appbuilder
marshmallow
marshmallow-sqlalchemy
aerospike
httplib2
pymssql
slackclient
SQLAlchemy
grpcio==1.70.0
grpcio-tools==1.70.0
python-keycloak-client
algoliasearch
confluent-kafka[avro,schemaregistry]==2.6.1
vertica-python
azure-mgmt-hdinsight
azure-identity
azure-storage-blob
alibabacloud-tea-openapi
alibabacloud-emr20160408
alibabacloud_ecs20140526
alibabacloud_vpc20160428
oss2
Flask-Bcrypt
azure-storage-blob
oauth2client
beautifulsoup4
dateparser
rocketchat-API
typeform
tableauserverclient
aiobotocore[boto3]
cachetools>=5.0.0,<6.0
apache-airflow-providers-databricks==6.5.0
apache-airflow-providers-snowflake==5.8.1
pytest
yapf==0.43.0
flake8==7.1.0
mypy==1.13.0
sumologic-sdk==0.1.14
types-protobuf
python-opsapi==2.0.174
pre-commit
uid2-client==2.5.18
slack-sdk
msgpack
msoffcrypto-tool
xlsxwriter
xlrd
opentelemetry-sdk==1.30.0
phonenumbers
airflow-clickhouse-plugin
databricks-sdk

